+++ pkg/factory/factory.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2025 Benjamin Borbe All rights reserved.
+// Copyright (c) 2026 Benjamin Borbe All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
@@ -27,4 +27,15 @@ func CreateSentryAlertHandler(sentryClient libsentry.Client) http.Handler {
 // Content-Type: application/json).
 func CreateHealthzHandler() http.Handler {
 	return handler.NewHealthzHandler()
+}
+
+// CreateObsidianRedirectHandler creates an HTTP handler that converts a
+// Telegram-carriable `https://` request into an `obsidian://open` deeplink.
+// `vaults` and `files` are the allowlist pair that constrains what the handler
+// is willing to emit; an empty allowlist allows nothing.
+func CreateObsidianRedirectHandler(
+	vaults handler.VaultAllowlist,
+	files handler.FileAllowlist,
+) http.Handler {
+	return handler.NewObsidianRedirectHandler(vaults, files)
 }